+++ packages/eyes-api/src/enums/IosDeviceName.ts
@@ -1,4 +1,7 @@
 export enum IosDeviceNameEnum {
+  iPhone_13_Pro_Max = 'iPhone 13 Pro Max',
+  iPhone_13_Pro = 'iPhone 13 Pro',
+  iPhone_13 = 'iPhone 13',
   iPhone_12_Pro_Max = 'iPhone 12 Pro Max',
   iPhone_12_Pro = 'iPhone 12 Pro',
   iPhone_12 = 'iPhone 12',
@@ -13,6 +16,7 @@ export enum IosDeviceNameEnum {
   iPhone_7 = 'iPhone 7',
   iPad_Pro_3 = 'iPad Pro (12.9-inch) (3rd generation)',
   iPad_7 = 'iPad (7th generation)',
+  iPad_9 = 'iPad (9th generation)',
   iPad_Air_2 = 'iPad Air (2nd generation)',
 }