+++ packages/eyes-sdk-core/lib/sdk/EyesVisualGrid.js
@@ -14,7 +14,7 @@ const EyesCore = require('./EyesCore')
 const CheckSettingsUtils = require('../sdk/CheckSettingsUtils')
 
 class EyesVisualGrid extends EyesCore {
-  static specialize({agentId, spec, VisualGridClient}) {
+  static specialize({agentId, spec, cwd, VisualGridClient}) {
     return class extends EyesVisualGrid {
       static get spec() {
         return spec
@@ -24,6 +24,9 @@ class EyesVisualGrid extends EyesCore {
       }
       get spec() {
         return spec
+      }
+      getCwd() {
+        return cwd
       }
       getBaseAgentId() {
         return agentId
@@ -59,7 +62,7 @@ class EyesVisualGrid extends EyesCore {
   async open(driver, optArg1, optArg2, optArg3, optArg4) {
     ArgumentGuard.notNull(driver, 'driver')
 
-    this._driver = await new Driver({spec: this.spec, driver, logger: this._logger._getNewLogger()}).init()
+    this._driver = await new Driver({spec: this.spec, driver, logger: this._logger}).init()
     this._context = this._driver.currentContext
 
     if (optArg1 instanceof Configuration) {
@@ -129,9 +132,7 @@ class EyesVisualGrid extends EyesCore {
   }
 
   async _check(checkSettings, closeAfterMatch = false, throwEx = true) {
-    this._logger.verbose(
-      `check started with tag "${checkSettings.name}" for test "${this._configuration.getTestName()}"`,
-    )
+    this._logger.log(`check started with tag "${checkSettings.name}" for test "${this._configuration.getTestName()}"`)
 
     return this._checkPrepare(checkSettings, async () => {
       const {persistedCheckSettings, cleanupPersistance} = await CheckSettingsUtils.toPersistedCheckSettings({
@@ -166,7 +167,7 @@ class EyesVisualGrid extends EyesCore {
           getEmulatedDevicesSizes: this._getEmulatedDevicesSizes,
           getIosDevicesSizes: this._getIosDevicesSizes,
           showLogs,
-          waitBeforeCapture: waitBeforeCapture,
+          waitBeforeCapture: () => utils.general.sleep(waitBeforeCapture),
         })
 
         const [{url}] = snapshots