+++ packages/driver/CHANGELOG.md
@@ -4,6 +4,11 @@
 
 - support cookies
 
+## 1.2.7 - 2021/10/30
+
+- updated to @applitools/types@1.0.18 (from 1.0.14)
+- updated to @applitools/utils@1.2.4 (from 1.2.3)
+
 ## 1.2.6 - 2021/10/7
 
 - fix issue with fractional viewport size on mobile devices