+++ packages/eyes-storybook/CHANGELOG.md
@@ -3,6 +3,30 @@
 ## Unreleased
 
 
+## 3.25.2 - 2021/10/30
+
+- replace legacy logger construction with new
+- deprecate `waitBeforeScreenshot` and introduce `waitBeforeCapture` instead
+- updated to @applitools/driver@1.2.7 (from 1.2.6)
+- updated to @applitools/eyes-puppeteer@1.8.5 (from 1.8.4)
+- updated to @applitools/eyes-sdk-core@12.23.24 (from 12.23.18)
+- updated to @applitools/test-server@1.0.6 (from 1.0.5)
+- updated to @applitools/visual-grid-client@15.8.43 (from 15.8.37)
+
+## 3.25.1 - 2021/10/20
+
+- updated to @applitools/eyes-puppeteer@1.8.4 (from 1.8.3)
+- updated to @applitools/eyes-sdk-core@12.23.18 (from 12.23.17)
+- updated to @applitools/visual-grid-client@15.8.37 (from 15.8.36)
+
+## 3.25.0 - 2021/10/18
+
+- updated to @applitools/driver@1.2.6 (from 1.1.3)
+- updated to @applitools/eyes-puppeteer@1.8.3 (from 1.8.2)
+- updated to @applitools/eyes-sdk-core@12.23.17 (from 12.23.0)
+- updated to @applitools/test-server@1.0.5 (from 1.0.4)
+- updated to @applitools/visual-grid-client@15.8.36 (from 15.8.21)
+
 ## 3.24.0 - 2021/8/23
 
 - add `startStorybookServerTimeout` config param