+++ packages/eyes-webdriverio-5/CHANGELOG.md
@@ -4,6 +4,18 @@
 
 - support cookies
 
+## 5.32.12 - 2021/10/20
+
+- updated to @applitools/dom-snapshot@4.5.10 (from 4.5.9)
+- updated to @applitools/eyes-sdk-core@12.23.18 (from 12.23.17)
+- updated to @applitools/visual-grid-client@15.8.37 (from 15.8.36)
+
+## 5.32.11 - 2021/10/20
+
+- updated to @applitools/dom-snapshot@4.5.9 (from 4.5.8)
+- updated to @applitools/eyes-sdk-core@12.23.17 (from 12.23.16)
+- updated to @applitools/visual-grid-client@15.8.36 (from 15.8.35)
+
 ## 5.32.10 - 2021/10/13
 
 - updated to @applitools/eyes-sdk-core@12.23.16 (from 12.23.15)