+++ packages/sdk-shared/CHANGELOG.md
@@ -5,6 +5,11 @@
 - support cookies
 - updated to @applitools/test-utils@1.0.7 (from 1.0.5)
 
+## 0.9.7 - 2021/10/30
+
+- updated to @applitools/test-server@1.0.6 (from 1.0.5)
+- updated to @applitools/test-utils@1.0.9 (from 1.0.7)
+
 ## 0.9.6 - 2021/9/10
 
 - updated to @applitools/test-server@1.0.5 (from 1.0.4)