+++ packages/utils/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## Unreleased
 
+- remove isGoogleFont method form utils and transfer it to rGridResource
+
+ ## 1.2.4 - 2021/10/27
+
+ - add a method to validate that a url is a google font resource
 
 ## 1.2.3 - 2021/9/9